Feature/nsalazar

+++ app/Query/Request/UserQuery.php
@@ -14,6 +14,7 @@ class UserQuery implements IUserQuery
     private $phone = 'phone';
     private $email = 'email';
     private $password = 'password';
+    private $rol_id = 'rol_id';
 
     public function index()
     {
@@ -34,6 +35,7 @@ class UserQuery implements IUserQuery
                     $this->phone     => 'required|numeric',
                     $this->email    => 'required|string|email|unique:users',
                     $this->password => 'required|string',
+                    $this->rol_id => 'required|numeric',
                 ]);
             } catch (\Exception $e) {
                 return response()->json(['message' => $e->getMessage()], 402);
@@ -45,10 +47,28 @@ class UserQuery implements IUserQuery
                 $this->phone     => $request->phone,
                 $this->email    => $request->email,
                 $this->password => bcrypt($request->password),
+                $this->rol_id => $request->rol_id,
             ]);
 
             $user->save();
             return response()->json(['message' => 'Usuario creado correctamente!'], 201);
+
+        } elseif (auth()->check() && auth()->user()->rol_id == 2) {
+            return response()->json(['message' => 'Rol de cliente, no tiene permisos de Admin!'], 201);
+
+        } elseif (auth()->check() && auth()->user()->rol_id == 3) {
+
+            $user = new User([
+                $this->name     => $request->name,
+                $this->lastname     => $request->lastname,
+                $this->phone     => $request->phone,
+                $this->email    => $request->email,
+                $this->password => bcrypt($request->password),
+                $this->rol_id => $request-> rol_id = 2,
+            ]);
+
+            $user->save();
+            return response()->json(['message' => 'Usurio creado con Rol cliente!'], 201);
         } else {
             return response()->json(['message' => 'No tiene permiso para crear usuarios!'], 403);
         }
@@ -79,20 +99,39 @@ class UserQuery implements IUserQuery
         return response()->json(['message' => 'User no exist!'], 404);
     }
 
-
-    /*  public function update(Request $request, Int $id)
+    public function update(Request $request, Int $id)
     {
-        return response()->json(['message' => 'Commerce update!'], 201);
-    }*/
+        if ($id > 0) {
+            $user = User::findOrFail($id);
+            if (is_null($user)) {
+                return response()->json(['message' => 'User no exist!'], 404);
+            } else {
+                $user->name = $request->name;
+                $user->lastname = $request->lastname;
+                $user->phone = $request->phone;
+                $user->email = $request->email;
+                $user->password = bcrypt($request->password);
+                $user->save();
+                return response()->json(['message' => 'User update!'], 201);
+            }
+        } else {
+            return response()->json(['message' => 'El id debe ser positivo!'], 404);
+        }
+    }
 
     public function destroy(Int $id)
     {
-        if ($id) {
-            $user = User::findOrFail($id);
-            $user->delete();
-            return response()->json(['message' => 'User destroy!'], 201);
+        if (auth()->check() && auth()->user()->rol_id == 1) {
+
+            if ($id) {
+                $user = User::findOrFail($id);
+                $user->delete();
+                return response()->json(['message' => 'User destroy!'], 201);
+            } else {
+                return response()->json(['message' => 'User no existe!'], 403);
+            }
         } else {
-            return response()->json(['message' => 'User no exist!'], 404);
+            return response()->json(['message' => 'No tiene permiso para Eliminar usuarios!'], 403);
         }
     }
 }
